.gitignore modified | Private scripts & plugins excluded
