+++ Plugins/SATBuildPhase/BuildPhasePlugin.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  StabilityAssuranceTool
+//
+//  Created by Andrii Sulimenko on 2025-01-09.
+//
+
+import Foundation
